hw: Change hw_get_time_us() to return a 64-bit time instead of 32-bit
A 32-bit counter for microsecond timestamps wraps around within about
an hour. This has the potential problem of causing hangs that are
virtually impossible to reproduce or debug.

+++ src/hw.h
@@ -49,6 +49,6 @@ extern void hw_audio_sfx_play(int sfx_index);
 extern void hw_audio_sfx_stop(void);
 extern bool hw_audio_sfx_volume(int volume/*0..128*/);
 
-extern uint32_t hw_get_time_us(void);
+extern int64_t hw_get_time_us(void);
 
 #endif
